test(security): isolate clawscan authority tests in dedicated file
